# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vũ Trụ Bóng Đá Ảo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  displayName?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  favoriteTeamId?: number | null;
  /** @nullable */
  favoriteTeamName?: string | null;
  /** @nullable */
  position?: string | null;
  rating?: number;
  matchesPlayed?: number;
  goals?: number;
  assists?: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfileUpdate {
  displayName?: string;
  avatarUrl?: string;
  position?: string;
}

export interface JoinTeamInput {
  teamId: number;
  position?: string;
}

export type LeagueType = typeof LeagueType[keyof typeof LeagueType];


export const LeagueType = {
  club: 'club',
  national: 'national',
} as const;

export interface League {
  id: number;
  name: string;
  shortName?: string;
  region: string;
  /** @nullable */
  country?: string | null;
  type: LeagueType;
  season: string;
  /** @nullable */
  logoUrl?: string | null;
  teamCount?: number;
  /** @nullable */
  description?: string | null;
}

export interface StandingRow {
  position: number;
  teamId: number;
  teamName: string;
  /** @nullable */
  teamLogoUrl?: string | null;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  /** @nullable */
  form?: string | null;
}

export interface TopScorer {
  playerId: number;
  playerName: string;
  teamId: number;
  teamName: string;
  nationality?: string;
  position?: string;
  goals: number;
  assists: number;
  matchesPlayed?: number;
}

export type TeamType = typeof TeamType[keyof typeof TeamType];


export const TeamType = {
  club: 'club',
  national: 'national',
} as const;

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  shortName?: string | null;
  region: string;
  country?: string;
  type: TeamType;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  stadiumName?: string | null;
  /** @nullable */
  founded?: number | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  secondaryColor?: string | null;
  playerCount?: number;
  /** @nullable */
  leagueId?: number | null;
  /** @nullable */
  leagueName?: string | null;
}

export type TeamDetailType = typeof TeamDetailType[keyof typeof TeamDetailType];


export const TeamDetailType = {
  club: 'club',
  national: 'national',
} as const;

export interface Player {
  id: number;
  name: string;
  /** @nullable */
  displayName?: string | null;
  nationality: string;
  position: string;
  /** @nullable */
  number?: number | null;
  /** @nullable */
  age?: number | null;
  teamId: number;
  teamName: string;
  rating?: number;
  pace?: number;
  shooting?: number;
  passing?: number;
  dribbling?: number;
  defending?: number;
  physical?: number;
  goals?: number;
  assists?: number;
  matchesPlayed?: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface Match {
  id: number;
  homeTeamId: number;
  homeTeamName: string;
  /** @nullable */
  homeTeamLogoUrl?: string | null;
  awayTeamId: number;
  awayTeamName: string;
  /** @nullable */
  awayTeamLogoUrl?: string | null;
  leagueId: number;
  leagueName: string;
  status: MatchStatus;
  scheduledAt: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  venue?: string | null;
}

export interface TeamDetail {
  id: number;
  name: string;
  /** @nullable */
  shortName?: string | null;
  region: string;
  country?: string;
  type: TeamDetailType;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  stadiumName?: string | null;
  /** @nullable */
  founded?: number | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  secondaryColor?: string | null;
  /** @nullable */
  leagueId?: number | null;
  /** @nullable */
  leagueName?: string | null;
  players?: Player[];
  recentMatches?: Match[];
}

export interface MatchInput {
  homeTeamId: number;
  awayTeamId: number;
  leagueId: number;
  scheduledAt: string;
}

export type MatchEventType = typeof MatchEventType[keyof typeof MatchEventType];


export const MatchEventType = {
  goal: 'goal',
  assist: 'assist',
  yellow_card: 'yellow_card',
  red_card: 'red_card',
  substitution: 'substitution',
  penalty: 'penalty',
  own_goal: 'own_goal',
} as const;

export interface MatchEvent {
  id: number;
  matchId: number;
  type: MatchEventType;
  minute: number;
  /** @nullable */
  playerId?: number | null;
  /** @nullable */
  playerName?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  description?: string | null;
}

export type MatchDetailStatus = typeof MatchDetailStatus[keyof typeof MatchDetailStatus];


export const MatchDetailStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface MatchDetail {
  id: number;
  homeTeamId: number;
  homeTeamName: string;
  /** @nullable */
  homeTeamLogoUrl?: string | null;
  awayTeamId: number;
  awayTeamName: string;
  /** @nullable */
  awayTeamLogoUrl?: string | null;
  leagueId: number;
  leagueName: string;
  status: MatchDetailStatus;
  scheduledAt: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  venue?: string | null;
  events?: MatchEvent[];
}

export interface StatsOverview {
  totalLeagues: number;
  totalTeams: number;
  totalPlayers: number;
  totalMatches: number;
  liveMatches: number;
  registeredUsers: number;
  /** @nullable */
  topLeague?: string | null;
  /** @nullable */
  topTeam?: string | null;
}

export type ListLeaguesParams = {
region?: string;
type?: ListLeaguesType;
};

export type ListLeaguesType = typeof ListLeaguesType[keyof typeof ListLeaguesType];


export const ListLeaguesType = {
  club: 'club',
  national: 'national',
} as const;

export type GetLeagueMatchesParams = {
status?: GetLeagueMatchesStatus;
};

export type GetLeagueMatchesStatus = typeof GetLeagueMatchesStatus[keyof typeof GetLeagueMatchesStatus];


export const GetLeagueMatchesStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export type ListTeamsParams = {
leagueId?: number;
region?: string;
type?: ListTeamsType;
search?: string;
};

export type ListTeamsType = typeof ListTeamsType[keyof typeof ListTeamsType];


export const ListTeamsType = {
  club: 'club',
  national: 'national',
} as const;

export type ListPlayersParams = {
teamId?: number;
position?: string;
nationality?: string;
search?: string;
};

export type ListMatchesParams = {
status?: ListMatchesStatus;
teamId?: number;
};

export type ListMatchesStatus = typeof ListMatchesStatus[keyof typeof ListMatchesStatus];


export const ListMatchesStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;