# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vũ Trụ Bóng Đá Ảo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Users
...
  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: List all leagues/tournaments
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [club, national]
      responses:
        "200":
          description: List of leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
  /leagues/{leagueId}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Get league details
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: League details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /leagues/{leagueId}/standings:
    get:
      operationId: getLeagueStandings
      tags: [leagues]
      summary: Get league standings table
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: League standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingRow"
  /leagues/{leagueId}/top-scorers:
    get:
      operationId: getLeagueTopScorers
      tags: [leagues]
      summary: Get top scorers in a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Top scorers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopScorer"
  /leagues/{leagueId}/matches:
    get:
      operationId: getLeagueMatches
      tags: [leagues]
      summary: Get matches for a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, finished]
      responses:
        "200":
          description: League matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"

  # Teams
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, email, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        favoriteTeamId:
          type: ["integer", "null"]
        favoriteTeamName:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        rating:
          type: number
        matchesPlayed:
          type: integer
        goals:
          type: integer
        assists:
          type: integer
        createdAt:
          type: string
...
    League:
      type: object
      required: [id, name, region, type, season]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        region:
          type: string
        country:
          type: ["string", "null"]
        type:
          type: string
          enum: [club, national]
        season:
          type: string
        logoUrl:
          type: ["string", "null"]
        teamCount:
          type: integer
        description:
          type: ["string", "null"]
    StandingRow:
      type: object
      required: [position, teamId, teamName, played, won, drawn, lost, goalsFor, goalsAgainst, goalDifference, points]
      properties:
        position:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        teamLogoUrl:
          type: ["string", "null"]
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        goalDifference:
          type: integer
        points:
          type: integer
        form:
          type: ["string", "null"]
    TopScorer:
      type: object
      required: [playerId, playerName, teamId, teamName, goals, assists]
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        teamId:
          type: integer
        teamName:
          type: string
        nationality:
          type: string
        position:
          type: string
        goals:
          type: integer
        assists:
          type: integer
        matchesPlayed:
          type: integer
...
    Match:
      type: object
      required: [id, homeTeamId, homeTeamName, awayTeamId, awayTeamName, leagueId, leagueName, status, scheduledAt]
      properties:
        id:
          type: integer
        homeTeamId:
          type: integer
        homeTeamName:
          type: string
        homeTeamLogoUrl:
          type: ["string", "null"]
        awayTeamId:
          type: integer
        awayTeamName:
          type: string
        awayTeamLogoUrl:
          type: ["string", "null"]
        leagueId:
          type: integer
        leagueName:
          type: string
        status:
          type: string
          enum: [upcoming, live, finished]
        scheduledAt:
          type: string
        homeScore:
          type: ["integer", "null"]
        awayScore:
          type: ["integer", "null"]
        minute:
          type: ["integer", "null"]
        venue:
          type: ["string", "null"]
...